Add Api s for exams to get the exams ans questions for all types os users

+++ Application/CQRS/Command/Exams/DeleteExamHandler.cs
@@ -1,5 +1,7 @@
 ﻿using Application.Common.Interfaces.CQRSInterfaces;
 using Application.Common.Interfaces.Presistance;
+using Domain.Enums;
+using Domain.Models;
 using Domain.Shared;
 
 namespace Application.CQRS.Command.Exams
@@ -17,6 +19,39 @@ namespace Application.CQRS.Command.Exams
         {
             try
             {
+                Exam examWillDeleted = await unitOfwork.ExamRepository.GetByIdAsync(request.ExamId);
+                if (examWillDeleted is null)
+                   return Result.Failure<int>(new Error(code: "DeleteExam", message: "No Exam has this Id"));
+
+                bool IfUserHasAccessToExam = false;
+
+                ExamPlace examPlaceOfCurrentExam = await unitOfwork.ExamPlaceRepository.GetByIdAsync(examWillDeleted.ExamPlaceId);
+                var user = await unitOfwork.UserRepository.GetUserByUserName(request.ExamCreatorUserName);
+
+                if (user is null)
+                    return Result.Failure<int>(new Error(code: "DeleteExam", message: "user has no access"));
+
+
+                if (examPlaceOfCurrentExam is null)
+                    return Result.Failure<int>(new Error(code: "DeleteExam", message: "can not delete this exam"));
+
+                if(examPlaceOfCurrentExam.ExamType==ExamType.Quiz && examPlaceOfCurrentExam.SectionId is not null)
+                {
+                    IfUserHasAccessToExam= await unitOfwork.SectionRepository.CheckIfInstructorInSection(InstrucotrId: user.Id, SectionId:(int) examPlaceOfCurrentExam.SectionId);
+                }
+                else if( (examPlaceOfCurrentExam.ExamType== ExamType.Quiz ||   examPlaceOfCurrentExam.ExamType == ExamType.Midterm) && examPlaceOfCurrentExam.CourseCycleId is not null)
+                {
+                    IfUserHasAccessToExam = await unitOfwork.CourseCycleRepository.CheckIfProfInCourseCycle(ProfId: user.Id,CourseCycleId: (int)examPlaceOfCurrentExam.CourseCycleId);
+                }
+                else if(examPlaceOfCurrentExam.ExamType==ExamType.Final || examPlaceOfCurrentExam.ExamType == ExamType.Semester)
+                {
+                    IfUserHasAccessToExam = true;
+                }
+
+                if(!IfUserHasAccessToExam)
+                    return Result.Failure<int>(new Error(code: "DeleteExam", message: "user has no access"));
+
+
                 bool Delted = await unitOfwork.ExamRepository.DeleteAsync(request.ExamId);
                 if (Delted)
                 {
